iq_formats: Add packed integer IQ formats
10 bit: 2 samples packed into 5 bytes
12 bit: 1 sample packed into 3 bytes
14 bit: 2 samples packed into 7 bytes

+++ src/qtgui/iq_tool.h
@@ -105,7 +105,7 @@ private:
 
     bool    is_recording;
     bool    is_playing;
-    int     bytes_per_sample;  /*!< Bytes per sample (fc = 4) */
+    float   bytes_per_sample;  /*!< Bytes per sample (fc = 4) */
     int     rec_bytes_per_sample;  /*!< Bytes per sample for recording */
     enum receiver::file_formats fmt;
     enum receiver::file_formats rec_fmt;
